+++ src/config/db/Barber.sql
@@ -447,3 +447,32 @@ INSERT INTO DATLICH VALUES(MADL_SEQ10.NEXTVAL, To_Date('24-12-2021','dd-mm-yyyy'
 INSERT INTO LoaiKhachHang VALUES (MALKH_SEQ2.NEXTVAL, 1, 'Vip', To_Date('11-09-2020','dd-mm-yyyy'), To_Date('02-09-2021','dd-mm-yyyy'))
 
 
+SELECT NHANVIEN.MANV,HO,TEN,NGAYSINH,GIOITINH,SODT,DIACHI,NGAYVAOLAM,LOAINHANVIEN,TINHTRANG,EMAIL,LUONG.MALUONG,LUONG.LUONGCOBAN,LUONGTHUONG,LUONGDUOCNHAN,NGAYNHANLUONG
+FROM NHANVIEN,LUONG,NHANLUONG
+WHERE NHANVIEN.MANV =LUONG.MANV
+AND    NHANVIEN.MANV = NHANLUONG.MANV
+
+SELECT MASP, TENSANPHAM, GIA, MOTASANPHAM, XUATXU, HINHANH, SANPHAM.TINHTRANG, SOLUONG, SANPHAM.MALSP, LOAISANPHAM.TENLOAISANPHAM FROM SANPHAM, LOAISANPHAM WHERE SANPHAM.MALSP = LOAISANPHAM.MALSP
+
+DESCRIBE HOADON
+INSERT INtO HOADON VALUES (MAHD_SEQ11.NEXTVAL,1,0,50000,1);
+INSERT INtO HOADON VALUES(MAHD_SEQ11.NEXTVAL,2,0,50000,1);
+INSERT INtO HOADON VALUES(MAHD_SEQ11.NEXTVAL,3,0,50000,0);
+SELECT * FROM CTHDSP
+DESCRIBE CTHDDV
+DESCRIBE CTHDSP
+INSERT INTO CTHDDV VALUES (1, 1);
+INSERT INTO CTHDDV VALUES (3, 1);
+INSERT INTO CTHDSP VALUES (2 ,3,1);
+INSERT INTO CTHDSP VALUES (2 ,4,1);
+
+SELECT HOADON.MAHD,MAKH,MADV,MASP,SOLUONG,KHUYENMAI,TONGTIEN,TINHTRANG 
+FROM HOADON,CTHDSP,CTHDDV
+WHERE  CTHDSP.MAHD =  CTHDDV.MAHD
+
+
+
+SELECT * FROM CTHDDV;
+
+ALTER TABLE NHANVIEN
+DROP COLUMN TEST;
